pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "financial-team-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "llama-cloud-services>=0.6.79",
    "llama-index-llms-openai>=0.6.7",
    "llama-index-workflows>=2.11.0",
    "resend>=2.19.0",
]

[tool.hatch.build.targets.wheel]
only-include = ["src/financial_team_agent"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.llamadeploy.workflows]
agent = "financial_team_agent.workflow:agent"

[tool.llamadeploy]
name = "financial-team"
env_files = [".env"]
llama_cloud = true
